follower manager types created
