+++ src/stores/follower-manager/useFollowerManagerStore.ts
@@ -1,22 +1,77 @@
-import {Session} from "next-auth";
+import {create} from "zustand";
+import axios from "axios";
+import {Follower, FollowerManagerState} from "@/stores/follower-manager/follower-manager.types";
 
-export interface Follower {
-    id: number;
-    login: string;
-    avatar_url: string;
-}
+export const useFollowerManager = create<FollowerManagerState>((set, get) => ({
+    followers: [],
+    following: [],
+    notFollowedBack: [],
+    selectedUsers: [],
+    loading: false,
 
-export interface FollowerManagerState {
-    followers: Follower[];
-    following: Follower[];
-    notFollowedBack: Follower[];
-    selectedUsers: string[];
-    loading: boolean;
     actions: {
-        fetchFollowersAndFollowing: (username: string, session: Session) => Promise<void>;
-        toggleSelection: (username: string) => void;
-        selectAll: () => void;
-        clearSelection: () => void;
-        unfollowSelected: (session: Session) => Promise<void>;
+
+        fetchFollowersAndFollowing: async (username: string, session) => {
+
+            set({loading: true});
+
+            try {
+                const [followersRes, followingRes] = await Promise.all([
+                    axios.get<Follower[]>(`https://api.github.com/users/${username}/followers`, {
+                        headers: {Authorization: `Bearer ${session}`},
+                    }),
+                    axios.get<Follower[]>(`https://api.github.com/users/${username}/following`, {
+                        headers: {Authorization: `Bearer ${session}`},
+                    }),
+                ]);
+
+                const followers = followersRes.data;
+                const following = followingRes.data;
+
+                const notFollowedBack = following.filter((followed) => !followers.some((follower) => follower.login === followed.login))
+
+                set({followers, following, notFollowedBack, loading: false});
+
+            } catch (error) {
+                console.error('Error fetching followers and following:', error);
+                set({loading: false});
+            }
+        },
+
+        toggleSelection: (username: string) => {
+            const {selectedUsers} = get();
+            if (selectedUsers.includes(username)) {
+                set({selectedUsers: selectedUsers.filter((user) => user !== username)});
+            } else {
+                set({selectedUsers: [...selectedUsers, username]});
+            }
+        },
+
+        selectAll: () => {
+            const {notFollowedBack} = get();
+            set({selectedUsers: notFollowedBack.map(user => user.login)});
+        },
+
+        clearSelection: () => {
+            set({selectedUsers: []});
+        },
+
+        unfollowSelected: async (session) => {
+            const {selectedUsers} = get();
+
+            try {
+                await Promise.all(
+                    selectedUsers.map((username) => {
+                        axios.delete(`https://api.github.com/user/following/${username}`, {
+                            headers: {Authorization: `Bearer ${session}`},
+                        })
+                    })
+                );
+                set({selectedUsers: []})
+            } catch (error) {
+                console.error('Error unfollowing users:', error);
+            }
+        }
+
     }
-}
+}))
